Create RTM.txt
Currently adding more to this. The Main Page GUI can be build #1. I'll be working on it this afternoon and evening til about midnight. 
I'm thinking we have this structure.

B1: GUI Main Page
B2: GUI and Functionality of Host Page
B3: GUI and Functionality of Client Page
B4: Main Page Functionality and Database Log Integration

Much of the GUI and  functionality for the host and client page are identical, if that is wrong please correct me.....lol. But we need 4 builds so I'm kinda slowing down the process so we have a complete 4 build project. Build 2 will likely be the most time consuming.